backend/requirements.txt
requests
playwright
python-dotenv
pyperclip
fastapi
uvicorn